PyCharm Srcs/Imports Fix + better .gitignore + requirements.txt
