+++ codes/data/LRHR_seg_bg_dataset.py
@@ -5,7 +5,7 @@ import numpy as np
 import torch
 import torch.utils.data as data
 
-import codes.data.util as util
+import data.util as util
 
 
 class LRHRSeg_BG_Dataset(data.Dataset):
