+++ codes/models/SRRaGAN.py
@@ -7,11 +7,12 @@ import torch
 import torch.nn as nn
 from torch.optim import lr_scheduler
 
-import codes.models.networks as networks
-from codes.models.modules.LPIPS import (
+import models.lr_schedulerR as lr_schedulerR
+import models.networks as networks
+from models.modules.LPIPS import (
     perceptual_loss as models,
 )  # import models.modules.LPIPS as models
-from codes.models.modules.loss import (
+from models.modules.loss import (
     GANLoss,
     GradientPenaltyLoss,
     HFENLoss,
@@ -21,8 +22,8 @@ from codes.models.modules.loss import (
     RelativeL1,
     L1CosineSim,
 )
-from codes.models.modules.losses import spl_loss as spl
-from codes.models.modules.losses.ssim2 import (
+from models.modules.losses import spl_loss as spl
+from models.modules.losses.ssim2 import (
     SSIM,
     MS_SSIM,
 )  # implementation for use with any PyTorch
@@ -31,7 +32,6 @@ from .__model__ import Model
 # from models.modules.losses.ssim3 import SSIM, MS_SSIM #for use of the PyTorch 1.1.1+ optimized implementation
 logger = logging.getLogger("base")
 
-import codes.models.lr_schedulerR as lr_schedulerR
 
 """ #debug
 def save_images(image, num_rep, sufix):
