+++ codes/test_abpn.py
@@ -7,11 +7,11 @@ from collections import OrderedDict
 import numpy as np
 import torch
 
-import codes.options.options as option
-import codes.utils.util as util
-from codes.data import create_dataset, create_dataloader
-from codes.data.util import bgr2ycbcr, patchify_tensor, recompose_tensor
-from codes.models import create_model
+import options as option
+import utils.util as util
+from data import create_dataset, create_dataloader
+from data.util import bgr2ycbcr, patchify_tensor, recompose_tensor
+from models import create_model
 
 
 # from torchvision.transforms import Compose, ToTensor
