+++ codes/test_ppon.py
@@ -4,11 +4,11 @@ import os
 import time
 from collections import OrderedDict
 
-import codes.options.options as option
-import codes.utils.util as util
-from codes.data import create_dataset, create_dataloader
-from codes.data.util import bgr2ycbcr
-from codes.models import create_model
+import options as option
+import utils.util as util
+from data import create_dataset, create_dataloader
+from data.util import bgr2ycbcr
+from models import create_model
 
 
 def main():
