+++ codes/test_seg.py
@@ -6,9 +6,9 @@ import numpy as np
 import torch
 import torchvision.utils
 
-import codes.models.modules.architectures.seg_arch as seg_arch
-import codes.utils.util as util
-from codes.data.util import imresize, modcrop
+import models.modules.architectures.seg_arch as seg_arch
+import utils.util as util
+from data.util import imresize, modcrop
 
 # options
 test_img_folder_name = "samples"  # image folder name
